requirements.txt
fabric
discord.py
python-dotenv
google-genai
matplotlib
pandas
